chore/#81: House 변수 타입 변경

+++ app/src/main/java/com/wearerommies/roomie/domain/entity/MoodCardEntity.kt
@@ -13,7 +13,7 @@ data class MoodCardEntity(
         val genderPolicy: String,
         val locationDescription: String,
         val isPinned: Boolean,
-        val contractTerm: Long,
+        val contractTerm: Int,
         val mainImgUrl: String,
     )
 }
